refactor with custom hook
