+++ src/components/NetworkGraph.tsx
@@ -158,68 +158,42 @@ const NetworkGraph = ({
       newNodes = newNodes.filter((node) => relatedNode.indexOf(node.id) > -1);
     }
 
-    const newGraph: GraphType = noEvent
-      ? {
-          graph: { nodes: newNodes, edges: newEdges },
-          events: {
-            hoverNode: () => {
-              document.body.style.cursor = 'pointer';
-            },
-            blurNode: () => {
-              document.body.style.cursor = 'default';
-            },
-            dragging: () => {
-              document.body.style.cursor = 'grabbing';
-            },
-            dragEnd: () => {
-              document.body.style.cursor = 'grab';
-              setTimeout(() => {
-                document.body.style.cursor = 'default';
-              }, 500);
-            },
-            zoom: ({ direction }: { direction: string }) => {
-              direction === '+'
-                ? (document.body.style.cursor = 'zoom-in')
-                : (document.body.style.cursor = 'zoom-out');
-              setTimeout(() => {
-                document.body.style.cursor = 'default';
-              }, 500);
-            },
-          },
-        }
-      : {
-          graph: { nodes: newNodes, edges: newEdges },
-          events: {
-            selectNode: ({ nodes }: { nodes: string[] }) => {
-              const noteId = nodes[0];
-              document.body.style.cursor = 'default';
-              navigate(`/note/${noteId}`);
-            },
-            hoverNode: () => {
-              document.body.style.cursor = 'pointer';
-            },
-            blurNode: () => {
-              document.body.style.cursor = 'default';
-            },
-            dragging: () => {
-              document.body.style.cursor = 'grabbing';
-            },
-            dragEnd: () => {
-              document.body.style.cursor = 'grab';
-              setTimeout(() => {
-                document.body.style.cursor = 'default';
-              }, 500);
-            },
-            zoom: ({ direction }: { direction: string }) => {
-              direction === '+'
-                ? (document.body.style.cursor = 'zoom-in')
-                : (document.body.style.cursor = 'zoom-out');
-              setTimeout(() => {
-                document.body.style.cursor = 'default';
-              }, 500);
-            },
-          },
-        };
+    const newGraph: GraphType = {
+      graph: { nodes: newNodes, edges: newEdges },
+      events: {
+        selectNode: ({ nodes }: { nodes: string[] }) => {
+          if (noEvent) {
+            return;
+          }
+          const noteId = nodes[0];
+          document.body.style.cursor = 'default';
+          navigate(`/note/${noteId}`);
+        },
+        hoverNode: () => {
+          document.body.style.cursor = 'pointer';
+        },
+        blurNode: () => {
+          document.body.style.cursor = 'default';
+        },
+        dragging: () => {
+          document.body.style.cursor = 'grabbing';
+        },
+        dragEnd: () => {
+          document.body.style.cursor = 'grab';
+          setTimeout(() => {
+            document.body.style.cursor = 'default';
+          }, 500);
+        },
+        zoom: ({ direction }: { direction: string }) => {
+          direction === '+'
+            ? (document.body.style.cursor = 'zoom-in')
+            : (document.body.style.cursor = 'zoom-out');
+          setTimeout(() => {
+            document.body.style.cursor = 'default';
+          }, 500);
+        },
+      },
+    };
     setState(newGraph);
   }, [id, filtBy, fontSizeNum, userNotes]);
