+++ src/pages/Editor/Editor.tsx
@@ -1,12 +1,11 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { v4 as uuidv4 } from 'uuid';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { NoteType } from '../../redux/types';
 import { addNote, editNote, selectProfile } from '../../redux/userSlice';
 import * as styles from '../../utils/styles';
 import CommentBox from './CommentBox';
 import EditorHelper from './EditorHelper';
+import { useEditNote } from './hooks';
 import './styles.css';
 
 const category = [
@@ -21,32 +20,11 @@ const category = [
 ];
 
 const Editor = () => {
-  const { id } = useParams();
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
+  const { id } = useParams();
   const profile = useAppSelector(selectProfile);
-  const [isEdit, setIsEdit] = useState(false);
-  const [note, setNote] = useState<NoteType>({
-    id: uuidv4(),
-    title: '',
-    content: '',
-    category: id === 'newNoteDevotion' ? 'devotion' : 'sermon',
-    link_notes: [],
-    create_time: new Date().getTime(),
-    edit_time: new Date().getTime(),
-  });
-
-  useEffect(() => {
-    if (!profile.isLogin) {
-      return;
-    }
-
-    if (id !== 'newNote' && id !== 'newNoteDevotion') {
-      const currentNote = profile.notes.find((note) => note.id === id);
-      setIsEdit(true);
-      setNote(currentNote!);
-    }
-  }, [profile.isLogin, isEdit]);
+  const { note, isEdit, setNote } = useEditNote(profile, id);
 
   const createLink = () => {
     const linkArray = [];
